requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
